Default Data and Zoom Course Graph
DATA$Default can be used to select which data is initially selected (manually set to the most recent currently)
Added a Course graph (Pulse vs Time in Video)
Brush on Course Graph allows one to zoom into the data on a second graph. Brush is on a reference version of the graph, which controls the second version of the graph.