More VCD and some Tweaks
More Viscera Cleanup Detail data and some tweaks to the Search.r script. The script now checks the modified time of the compressed CSV against the oldest of the CSVs in the folder to determine if it should combine the two objects. This is really only for VCD recordings, but should be fine for any other.