Update TBOG - Archive Sizes.r
Realized it needed the ability to better handle disagreement between the CSV and new information. This will now check for missing values (NA) and replace them with values.
This could also be achieved with similar dplyr functions, but as those functions tend to replicate base functions with different names and slightly different formatting, I prefer to use base R functions. This way the library does not need to be loaded.
Also have it get the file size differently, so now it keeps the sizes paired with the file name, in case there are alignment issues.