Graph Download module and fixed caching
The graph downloading stuff is now in modules, which is appropriate
Also fixed up caching as before it would associate cached items with the wrong selections
There is a problem with it not completely refreshing graphs when changing the loaded data, but selecting a different part and coming back fixes it. I truly do not understand why this is happening, but it is.